project/api/urls.py
from django.urls import include, path
from rest_framework import routers
from . import views
from rest_framework.urlpatterns import format_suffix_patterns

router = routers.DefaultRouter()
router.register(r'users', views.UserViewSet)

# Wire up our API using automatic URL routing.
# Additionally, we include login URLs for the browsable API.
urlpatterns = [
    path('', include(router.urls)),
    path('api-auth/', include('rest_framework.urls', namespace='rest_framework')),
    path('create/user/', views.CreateUser().as_view()),
    path('user/<int:pk>', views.GetUser.as_view()),
    path('update/user/<int:pk>', views.UpdateUser.as_view()),
]